Fix exception objects not being logged properly
The order of arguments sent to `_logger.Error` in
`ExceptionHandlingMiddleware` was incorrect, causing exception objects
to not be logged properly. The exception object was passed in as a
format parameter, but as the log message didn't contain any formatting
placeholders, the exception object was lost in the resulting log output.

This change puts the exception object as the first argument, as the
Error log method requires.